requirements.txt
google-api-python-client==2.26.1
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
httplib2==0.19.1
oauth2client==4.1.3
oauthlib==3.1.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
tabulate==0.8.9
termcolor==1.1.0
uritemplate==4.1.1
